[IMP] point_of_sale: rewrite price access control logic
Introduced at 3a4fc7111b6d83503015c0118fe54e0ffefe7af1. Instead of
doing hacky things like self.pos.gui.screen_instances... in the
cashier widget to force the numpad UI to rerender itself it's better
to use Backbone.js because:

1. we're already using it everywhere else to handle UI updates
2. easier to extend for eg. adding more cashier related access rights

To do so cashier is turned into a Backbone attribute on PosModel. All
references to the cashier property are replaced with get_cashier to
ensure the Backbone attribute is used.

Note that this is not necessary for the user property on
PosModel. This contains the actual user logged into Odoo and thus
never changes.

This also removes the check in set_unit_price because:

1. it was wrong to begin with (no price change when
   restrict_price_control is enabled or you're a manager?)
2. even if fixed it would break the following for non-managers:
   1. tipping
   2. fix_tax_included_price