[FIX] website_sale(_delivery): limit 'Setup payment / delivery methods' to group_system
When choosing a payment acquirer or a delivery provider there is a link
allowing to directly jump in the backend to configure them. It has been
introduced at 0e6912ffc70566fc960428696519cbe5e6098ee0.

However it is displayed for every user going through the checkout process.
This commit limits it to group_system. It seems a better idea to not show
it to everyone as they don't have access to it anyway.